requirements.txt
jinja2==3.1.2

sortedcontainers==2.4.0
PyYAML==6.0.1

#torch-ac>=1.4.0
minigrid>=2.2.0
tensorboardX>=1.6
numpy>=1.3
gymnasium>=0.26
pygame>=2.5.2
array2gif>=1.0.4
